Enhances Cloud Run deploy and cleanup workflows (#15)
* Adds support for docker buildx bake

Introduces the ability to use `docker buildx bake`
as an alternative build tool, allowing for more
complex build configurations. This includes the
ability to specify multiple Dockerfiles and target
specific build targets within a bake file.

* Add Docker Buildx setup step in CI workflow

This commit introduces a new step to set up Docker Buildx in the GitHub Actions workflow, enhancing the build process by allowing the use of the Buildx driver for more efficient Docker image builds.

* Adds custom command and args to Cloud Run deploy

Enables users to specify custom command and arguments for the Cloud Run service deployment via the configuration file.

This change extracts the `command` and `args` from the config file, formats them, and passes them as flags to the `gcloud run deploy` command.

* Adjusts command and args parsing for Cloud Run

Parses the command arguments from the config file using space separation instead of comma separation.

Ensures that arguments passed to the Cloud Run service are correctly quoted, allowing for proper handling of spaces and special characters within the arguments.

* Adds support for app config file via base64

Enables the passing of an application configuration file to the Cloud Run service as a base64 encoded environment variable.

This simplifies configuration management by allowing a single configuration file to be used across different environments without requiring modifications to the container image.

* Adds health check probes to Cloud Run deployments

Implements startup and liveness probe configuration for Cloud Run deployments using values defined in the application config file.

This allows for more robust and reliable deployments by enabling Cloud Run to automatically check the health of the service and restart it if necessary. Supports both HTTP and gRPC probes.

* Adds debug capabilities to Cloud Run deploy

Adds debugging steps to the Cloud Run deploy workflow for troubleshooting.

This includes printing the raw config file, parsed config values, environment variables, and the contents of the GITHUB_OUTPUT file. These steps aid in diagnosing issues related to configuration and deployment.

* Improves health check probe handling

Updates the health check probe logic to correctly determine the presence of startup and liveness probes in the Cloud Run deployment configuration.
It uses a more robust method to check for probe definitions, preventing potential issues with empty or missing probe configurations.
Also, removes debugging steps and prints.

* Adds support for specifying a port

Allows specifying the port for the Cloud Run service
through the configuration.

This commit introduces a new flag that enables users to define
the port on which their service listens.

* Exposes service port from config

Allows specifying the service port in the configuration file.
This enables greater flexibility in defining service parameters.

* Appends issue number to final service name

Ensures unique service names across multiple environments by appending the GitHub issue number when a config file is used.
This addresses potential naming conflicts.

* Adds config file option for service name

Allows specifying a configuration file to determine the Cloud Run service name, enhancing flexibility.
Removes the direct service name and branch-based name input options.
The service name is now derived from the config file or the service name input parameter.
If a config file is present, reads the service name from it; otherwise, defaults to the provided service name.
The service name is also postfixed with the pull request number on pull requests.

* Allows unauthenticated access

Adds a step to explicitly allow unauthenticated access to the Cloud Run service if the `allow_unauth` configuration option is set to true.

This ensures that services intended to be publicly accessible do not require authentication, simplifying access for users.

* Removes public access step

Removes the step that explicitly allows unauthenticated access to the Cloud Run service. This is likely being handled elsewhere or is no longer necessary.

* Replaces github event number in app config

Ensures that the `${{ github.event.number }}` placeholder within the application configuration file is correctly substituted with the actual event number before base64 encoding. This allows dynamic configuration based on the GitHub event that triggered the workflow.

* Fixes regex for app config replacement

The previous regex was not correctly escaping the dollar sign, causing the variable replacement to fail.

This commit updates the regex to properly escape the dollar sign, ensuring that the correct value is replaced in the app config file.

* Replaces github event number in app config

Updates the app config file replacement logic to correctly substitute the github event number.

The previous implementation was not correctly replacing the variable. This change ensures that the substitution works as expected.

* Updates service name generation

Modifies the service name generation logic to only append the pull request number when the event is a pull request.
This prevents appending the pull request number to service names in other workflows.