Allows the workflow to configure access to private Go modules. (#33)
* Fix _checks-golang

* Adds support for private Go modules

Allows the workflow to configure access to private Go modules.
It adds an input parameter `goprivate` to specify the module path prefixes.
It also configures Git to use the GitHub App token for authentication
when accessing private repositories.

* _atlas-migratios: Enables checkout using GitHub App authentication

Adds support for authenticating with a GitHub App to access private repositories.
This allows the workflow to checkout code and configure Git using the GitHub App's token,
enabling access to repositories without relying on user-specific credentials.

Also adds support for configuring GOPRIVATE environment variable.

* Enables GitHub App authentication for private repos

Adds support for authenticating with a GitHub App to access private repositories during the build process.
This allows the workflow to clone private repositories and use their contents during the Docker image build.
It configures Git to use a GitHub App token for authentication and passes the token as a build argument.

* minor changes

* minor change

* minor changes

* minor change

* test actions

* changes

* test new branch

* test on checks-golang

* use @main actions